feat: virtualisation
Added the ability to switch between a setup using the GPU and a
passthrough for VMs. Do note that while the passthrough works, I have
not figured out how to:
A) Reset the card upon VM reset. So the GPU is in a locked state where
it won't reset, and therefore the VM can't get it the next time it boots
up. I have no figured out a solution yet.
B) Maybe a setting is wrong. I have not gotten GPU passthrough to work
correctly with Windows VM. I can get the bios screen to show up, but
post boot shows a messed up output. I have no clue why...