feat: emacs lsp and custom dashboard
I finally setup the lsp for emacs.
I have not setup the quickshell/qml for emacs as of yet. It requires an
emacs overlay and some other things, which I'd rather have in a seperate
commit.

There are warnings when opening nixos folder as an entire project. It
seems it can't pull the packages and confirm their hashes. I'll have to
fix that.