feat: automatic user generation
Now I can add users and have them picked up automatically, just by
making new folders. Each user can have their own programs.

The users are declared in the configuration of the machines.

This makes it so that for example a server setup does not have a bunch
of un-needed user programs.