feat: host context
This is in preparation for my shared user between linux/Darwin.
Note that hjem does not support Darwin yet because of systemd...
This also allows me to describe if a setup is headless or not.
This means I could probably also make my own user shared between my
workstation and server, because I can gaurd any of the new program
bundles with a context.